Do not export rsaKeySizeBits constant
After discussion with other TAs, we cannot think of any
reason that students would need to use this constant since
PKEKeyGen() and DSKeyGen() have the constant hardcoded already.

Therefore, we are renaming the constant to be semantically
aligned with rsa.GenerateKey(), which accepts the key size
in bits, and also make it start with a lower-case letter
so that it is private to the module (i.e. students cannot
reference this constant in their code).